# backend/docs/swagger.yaml
basePath: /
definitions:
  authModel.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  authModel.ResponseToken:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  bankModel.Bank:
    properties:
      id:
        type: string
      kode_bank:
        type: string
      nama:
        type: string
    type: object
...
  bankModel.BankUpdate:
    properties:
      kode_bank:
        type: string
      nama:
        type: string
    type: object
  katalogModel.KatalogCreate:
    properties:
      harga:
        type: integer
      produk_id:
        type: string
      user_id:
        type: string
    type: object
  katalogModel.KatalogShow:
    properties:
      harga_katalog:
        type: integer
      id:
        type: string
      produk:
        $ref: "#/definitions/produkModel.ProdukShow"
    type: object
  katalogModel.UpdateKatalog:
    properties:
      harga:
        type: integer
    type: object
...
  kategoriModel.KategoriReference:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori_id:
        type: string
    type: object
  kategoriModel.KategoriShow:
    properties:
      deskripsi:
        type: string
      id:
        type: string
      nama:
        type: string
      ref_kategori:
        $ref: "#/definitions/kategoriModel.RefKategoriShow"
    type: object
...
  kategoriModel.RefKategoriShow:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
  kategoriModel.Reference:
    properties:
      id:
        type: string
      nama:
        type: string
      values:
        items:
          $ref: "#/definitions/kategoriModel.KategoriReference"
        type: array
    type: object
  model.KatalogTest:
    properties:
      meta:
        $ref: "#/definitions/model.PaginationMeta"
      result:
        items:
          $ref: "#/definitions/katalogModel.KatalogShow"
        type: array
    type: object
...
  model.PaginationMeta:
    properties:
      next:
        type: string
      page:
        type: integer
      previous:
        type: string
      reference:
        $ref: "#/definitions/model.ReferenceResult"
      size:
        type: integer
      total_data:
        type: integer
    type: object
...
  model.ReferenceResult:
    properties:
      data:
        items:
          $ref: "#/definitions/kategoriModel.Reference"
        type: array
      last:
        type: string
    type: object
  model.ResponseMessage:
    properties:
      message:
        type: string
      reference:
        $ref: "#/definitions/model.ReferenceResult"
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  produkModel.ProdukShow:
    properties:
      detail:
        type: string
      foto:
        type: string
      harga:
        type: integer
      id:
        type: string
      kategori:
        $ref: "#/definitions/kategoriModel.KategoriShow"
      nama:
        type: string
    type: object
...
  userModel.UserCreate:
    properties:
      email:
        type: string
      foto:
        type: string
      nama:
        type: string
      no_telp:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: dokumentasi API Kavi Kasir
  title: KAVI Kasir API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
        - application/json
      description: Login
      parameters:
        - description: Field login
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/authModel.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/authModel.ResponseToken"
      summary: Login
      tags:
        - auth
  /auth/register:
    post:
      consumes:
        - application/json
      description: Membuat akun
      parameters:
        - description: Field order
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/userModel.UserCreate"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Membuat akun
      tags:
        - auth
  /bank:
    post:
      consumes:
        - application/json
      description: Tambah data bank
      parameters:
        - description: Field produk
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/bankModel.Bank"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/bankModel.Bank"
      summary: Tambah data bank
      tags:
        - bank
  /bank/{bank_id}:
    delete:
      consumes:
        - application/json
      description: Hapus data bank user
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Hapus data bank user
      tags:
        - bank
    put:
      consumes:
        - application/json
      description: Mengubah data bank
      parameters:
        - description: bank_id
          in: path
          name: bank_id
          required: true
          type: string
        - description: Field bank
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/bankModel.BankUpdate"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Ubah bank
      tags:
        - bank
...
  /katalog:
    get:
      consumes:
        - application/json
      description: Menampilkan data katalog
      parameters:
        - description: page
          in: path
          name: page
          type: integer
        - description: size
          in: path
          name: size
          type: integer
        - description: keyword
          in: path
          name: keyword
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.KatalogTest"
      summary: Menampilkan data katalog
      tags:
        - katalog
    post:
      consumes:
        - application/json
      description: Menambahkan katalog
      parameters:
        - description: Field order
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/katalogModel.KatalogCreate"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Menambahkan katalog
      tags:
        - katalog
  /katalog/{katalog_id}:
    delete:
      consumes:
        - application/json
      description: Menghapus data katalog
      parameters:
        - description: katalog_id
          in: path
          name: katalog_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/model.ResponseMessage"
      summary: Hapus katalog
      tags:
        - katalog
    get:
      consumes:
        - application/json
      description: ambil data katalog
      parameters:
        - description: katalog_id
          in: path
          name: katalog_id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/katalogModel.KatalogShow"
      summary: Ambil katalog berdasarkan id
      tags:
        - katalog
    patch:
      consumes:
        - application/json
      description: Mengubah harga katalog
      parameters:
        - description: katalog_id
          in: path
          name: katalog_id
          required: true
          type: string
        - description: Field katalog
          in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/katalogModel.UpdateKatalog"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/katalogModel.KatalogShow"
      summary: Ubah harga katalog
      tags:
        - katalog
...
swagger: "2.0"